Complete first draft of gain graph implementation
This commit also modifies conf.py to use autodoc_default_options instead
of napoleon_include_special_with_doc to include special class and module
members (i.e. those whose names start and end with "__"). This is
because the Napoleon option overrides any autodoc options that would
otherwise exclude these members. "__init__" and "__weakref__" have been
set to be excluded by default, as they were excluded before, but this
behavior (and that of including special members) can be modified on a
per-module basis, which was not possible before.

This became an issue with GainGraph, whose documentation includes
members from :inherited-members: but excludes members inherited from
NetworkX classes for parity with pyrigi.Graph. The reason pyrigi.Graph
does not trigger this issue is that it does not set :inherited-members:,
so the return value of autodoc._member_finder._get_members_to_document
does not include members of NetworkX classes, which prevents Napoleon
from receiving the relevant autodoc-skip-member events and forcing
these members to be included.